move ParameterizedQuery to models folder
