+++ redash/models/parameterized_query.py
